solana-rfp-app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
openai>=1.3.0
pypdf2>=3.0.0
python-docx>=1.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
email-validator>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
